Fix, restyle, implementazioni varie da revisione pre-rilascio

+++ src/app/modules/accountant/components/accountant-events-list/accountant-events-list.component.ts
@@ -2,8 +2,14 @@ import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
 
+// Componenti
+import { ConfirmModalComponent } from '../../../../components';
+
+// Moduli Esterni
+import { BsModalService } from 'ngx-bootstrap/modal';
+
 // Services
-import { EventService } from '../../../../services';
+import { EventService, CharacterService } from '../../../../services';
 
 // Models
 import {
@@ -11,6 +17,10 @@ import {
   GameEventSubscriptionDTO,
   RestResponse
 } from '../../../../models';
+
+// Utility
+import { FileUtils } from '../../../../utils/fileUtils';
+const FileSaver = require('file-saver');
 
 @Component({
   selector: 'app-accountant-events-list',
@@ -26,8 +36,10 @@ export class AccountantEventsListComponent implements OnInit {
 
   constructor(
     private eventService: EventService,
+    private characterService: CharacterService,
     private router: Router,
-    private activatedRoute: ActivatedRoute
+    private activatedRoute: ActivatedRoute,
+    private bsModalService: BsModalService
   ) {}
 
   ngOnInit() {
@@ -52,6 +64,15 @@ export class AccountantEventsListComponent implements OnInit {
           this.oldEventsList = [...this.oldEventsList, eventItem];
         }
       });
+
+      this.oldEventsList = this.oldEventsList.sort(
+        (event1: GameEvent, event2: GameEvent) => {
+          return (
+            new Date(event2.datainizio).getTime() -
+            new Date(event1.datainizio).getTime()
+          );
+        }
+      );
     });
   }
 
@@ -102,10 +123,84 @@ export class AccountantEventsListComponent implements OnInit {
     });
   }
 
+  public subscriptionHasCharacters(
+    eventSubscription: GameEventSubscriptionDTO
+  ): boolean {
+    return (
+      eventSubscription &&
+      eventSubscription.dettagli.some(dettaglio => !!dettaglio.idpersonaggio)
+    );
+  }
+
+  public printCharacterSheet(
+    eventSubscription: GameEventSubscriptionDTO
+  ): void {
+    const characterIDs = eventSubscription.dettagli
+      .filter(dettaglio => !!dettaglio.idpersonaggio)
+      .map(dettaglio => dettaglio.idpersonaggio)
+      .filter((elem, index, self) => index === self.indexOf(elem));
+
+    this.characterService.PrintCharactersSheets(characterIDs).subscribe(res => {
+      if (!res.payload) {
+        return;
+      }
+
+      const blob = FileUtils.b64toBlob(res.payload, 'application/zip');
+      FileSaver.saveAs(blob, 'characters-sheets.zip');
+    });
+  }
+
+  public printEventCharacterSheet(eventData: GameEvent, onlyPaid = true): void {
+    let eventSubscriptions = this.eventSubscriptions[`key${eventData.id}`];
+
+    if (onlyPaid) {
+      eventSubscriptions = eventSubscriptions.filter(
+        subscription => subscription.pagata
+      );
+    }
+
+    const characterIDs: number[] = [];
+    eventSubscriptions.forEach(subscription => {
+      const ids = subscription.dettagli
+        .filter(dettaglio => !!dettaglio.idpersonaggio)
+        .map(dettaglio => dettaglio.idpersonaggio)
+        .filter((elem, index, self) => index === self.indexOf(elem));
+
+      ids.forEach(characterId => {
+        if (characterIDs.indexOf(characterId) === -1) {
+          characterIDs.push(characterId);
+        }
+      });
+    });
+
+    if (!characterIDs || (characterIDs && characterIDs.length === 0)) {
+      return;
+    }
+
+    this.characterService.PrintCharactersSheets(characterIDs).subscribe(res => {
+      if (!res.payload) {
+        return;
+      }
+
+      const blob = FileUtils.b64toBlob(res.payload, 'application/zip');
+      FileSaver.saveAs(blob, 'characters-sheets.zip');
+    });
+  }
+
   public deleteSubscription(eventId: number, subscriptionId: number): void {
-    this.eventService.RemoveSubscription(subscriptionId).subscribe(res => {
-      if (res.payload) {
-        this.loadEventSubscriptions(eventId);
+    const modalRef = this.bsModalService.show(ConfirmModalComponent);
+    modalRef.content.title = 'Conferma eliminazione';
+    modalRef.content.content = `Confermare l'eliminazione dell'iscrizione?`;
+    modalRef.content.closebutton = `Annulla`;
+    modalRef.content.okbutton = `Conferma`;
+
+    this.bsModalService.onHide.subscribe(() => {
+      if (modalRef.content.result) {
+        this.eventService.RemoveSubscription(subscriptionId).subscribe(res => {
+          if (res.payload) {
+            this.loadEventSubscriptions(eventId);
+          }
+        });
       }
     });
   }
